requirements.txt
Flask==3.0.0
flask-cors==4.0.0
yt-dlp>=2024.1.0
requests>=2.31.0
urllib3>=2.0.0
gunicorn>=20.1.0
Flask==3.0.0
flask-cors==4.0.0
yt-dlp>=2024.1.0
requests>=2.31.0
urllib3>=2.0.0

pymongo
dnspython
supabase
python-dotenv